force common dtype for input vars in bias correction

+++ nlsam/tests/test_bias_correction.py
@@ -6,9 +6,9 @@ from nlsam.bias_correction import root_finder_sigma, multiprocess_stabilization
 
 def test_root_finder_sigma():
 
-    eta = np.full((10, 10, 10, 10), 100.)
-    sigma = np.full((10, 10, 10, 10), 1.)
-    N = np.full((10, 10, 10, 10), 1.)
+    eta = np.full((10, 10, 10, 10), 100., dtype=np.float32)
+    sigma = np.full((10, 10, 10, 10), 1., dtype=np.float32)
+    N = np.full((10, 10, 10, 10), 1., dtype=np.float32)
     mask = np.ones_like(eta[..., 0])
     output = root_finder_sigma(eta, sigma, N, mask)
     assert_allclose(output, sigma, atol=1e-4)
@@ -23,8 +23,8 @@ def test_root_finder_sigma():
 
     # magnitude SNR of 0.5 -> sigma = 50, eta = 25 for N = 1
     a = [np.sqrt((25 + np.random.randn(10000) * 50)**2 + (np.random.randn(10000) * 50)**2) for ii in range(250)]
-    eta = np.mean(a, axis=1, keepdims=True)
-    sigma = np.std(a, axis=1, keepdims=True)
+    eta = np.mean(a, axis=1, keepdims=True, dtype=np.float32)
+    sigma = np.std(a, axis=1, keepdims=True, dtype=np.float64)
     N = np.ones_like(sigma)
     output = root_finder_sigma(eta, sigma, N)
